Don't tag dicts for Semantria responses
I speculated that having tags would be valuable.  I don't actually have
a use case for it yet, so leave the tag as a variable for now.  As a
consequence, I removed the one dependency on library(interpolate).
Using it would have required some singleton hacks that aren't worth it.