Raise a condition on request/3 error
If request/3 encounters an error while making its HTTP request to
Semantria, raise a signal (via library(condition)).  This allows the
application that's consuming Semantria's API to decide how these kinds
of failures should be handled.  In many cases, a simple fail is enough.
In other cases, an exception is wanted.